requirements-old.txt
absl-py==1.2.0
aiosignal==1.2.0
ale-py==0.7.5
asttokens==2.0.8
attrs==22.1.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
backcall==0.2.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.4
cloudpickle==2.1.0
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
distlib==0.3.6
dm-control==1.0.5
dm-env==1.5
dm-tree==0.1.7
entrypoints==0.4
executing==1.0.0
fasteners==0.18
filelock==3.8.0
fonttools==4.37.1
frozenlist==1.3.1
glfw==2.5.4
grpcio==1.43.0
gym==0.20.0
gym-notices==0.0.8
gym-robotics==1.0.1
gym-unity==0.28.0
Gymnasium==0.26.3
gymnasium-notices==0.0.1
idna==3.3
imageio==2.21.3
importlib-metadata==4.13.0
importlib-resources==5.9.0
ipykernel==6.15.2
ipython==8.5.0
jedi==0.18.1
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
labmaze==1.0.5
lxml==4.9.1
lz4==4.0.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mlagents-envs==0.28.0
msgpack==1.0.4
mujoco==2.2.2
mujoco-py==2.1.2.14
nest-asyncio==1.5.5
networkx==2.8.6
numpy==1.21.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.6.0.66
packaging==21.3
panda-gym==3.0.0
pandas==1.4.3
parso==0.8.3
PettingZoo==1.15.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
prompt-toolkit==3.0.31
protobuf==3.20.1
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pybullet==3.2.5
pycparser==2.21
pyglet==1.5.27
Pygments==2.13.0
PyOpenGL==3.1.6
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.0
PyYAML==6.0
pyzmq==23.2.1
ray==2.0.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.1
seaborn==0.12.0
six==1.16.0
stable-baselines3==1.6.2
stack-data==0.5.0
style==1.1.0
tabulate==0.8.10
tensorboardX==2.5.1
tifffile==2022.8.12
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
typing_extensions==4.3.0
update==0.0.1
urllib3==1.26.12
virtualenv==20.16.5
wcwidth==0.2.5
zipp==3.8.1
hydra-core==1.3.1
omegaconf==2.3.0